Number and string assignment
You learned how to create variables and assign data to them earlier.

This time, let's use it by assigning numbers and strings to variables through practice.

Type conversion is essential to operate on variables of different data types or to change the format in which data is expressed.

Where did the type conversion take place in this lab? Take a good look!


Let's do this!
Take a look at the skeleton code written in the code editor, and understand how the code works.
Try assigning values ​​to the heartbeats and name variables in lines 2 and 15 by entering a number and a string, respectively.
Click the Run button to check the result.
Tips
Take a look at the print() function on the last line! Can you see the addition (+) operation between strings and strings?

In addition, the heartbeats variable, which is an integer type, is converted into a string data type using the str() function and operation is being performed!